+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project are documented in this file.
 ## 0.3.0 - ????-??-??
 
 - Improved documentation formatting.
+- Rename mockfn.macros to mockfn.plain.
 
 ## 0.2.0 - 2025-02-18